[FEAT] market에 상품 없을 때 null 처리

+++ src/main/java/com/example/letsGo/controller/market/ProductDetailController.java
@@ -35,7 +35,7 @@ public class ProductDetailController {
     }
     @GetMapping("/detail")
     public String getDetailProduct(
-            @RequestParam("productId") int productId,
+            @RequestParam("productId") Long productId,
             @RequestParam("productType") int productType,
             Model model
     ) {
